Fix errors with assert in tests under php8
Since PHP 7, `assert()` would throw an `AssertionError` if the ini
option `assert.exception` was enabled, but it didn't become the default
until PHP 8 so instead warnings were converted to an `ErrorException`.

This commit updates phpunit's bootstrap script to make sure that the
`assert.exception` option is enabled for both PHP 7 *and* 8, enabling us
to switch the expectation to a proper `AssertionError` in the tests.