Update phpunit and infection php
Bumps infection, phpunit and php-invoker to the latest versions to get
support for php8 and the xdebug config changes.

With the update to phpunit comes a new xml config schema. The new
`coverage` section is made up of the `filter.whitelist` value (now
'include') and the `log.coverage-*` values as children of `report`.