Add a test to cover mutation in CollectionTuple
After updating the upstream image to include pcntl, Infection PHP
flagged a mutation in `Collection\CollectionTuple::__toArray()` where
`$this->value instanceof ShellInterface` could be changed to `false`.

That results in the first half of the ternary never executing. To catch
it, this commit adds a test for `__toArray()` that uses an instance of
`ShellBuilder`, complimenting the existing string-based test.