Switch to using dual PHPUnit configs
Phpunit 9 won't run under 7.2, but phpunit 8 doesn't understand the
config for v9. This here is an attempt at a "dual boot" of sorts.

The old config is restored to a separate file so we can pass the legacy
config for 7.2 but still use the new config format for other versions.