Enable xdebug coverage mode for php8 in travis
For php 8, we need to manually set the new `xdebug.mode` option to
enable coverage mode, but it won't work if we simply do something like
`php -d xdebug.mode=coverage composer infection` because InfectionPHP
spawns new processes that would lose that option.

To provide for this, there's the `--initial-tests-php-options` shell
option to infection that will pass that on when it calls PHP for the
initial test run, which is the only time we need coverage enabled.

Since this is only required in PHP 8 and there are no other parts using
the `PHPUNIT_LEGACY` env variable, we can instead test against php 7.x
to remove the need for PHPUNIT_LEGACY entirely.